requirements.txt
annotated-types==0.6.0
anyio==4.2.0
boto3==1.34.8
botocore==1.34.8
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fastapi==0.108.0
filelock==3.13.1
fsspec==2023.12.2
h11==0.14.0
huggingface-hub==0.20.1
idna==3.6
imageio==2.33.1
Jinja2==3.1.2
jmespath==1.0.1
lazy_loader==0.3
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.2
opencv-python-headless==4.8.1.78
packaging==23.2
pdf2image==1.16.3
Pillow==10.1.0
pip==23.2.1
pyclipper==1.3.0.post5
pydantic==2.5.3
pydantic_core==2.14.6
python-bidi==0.4.2
python-dateutil==2.8.2
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
s3transfer==0.10.0
safetensors==0.4.1
scikit-image==0.22.0
scipy==1.11.4
setuptools==65.5.0
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
starlette==0.32.0.post1
sympy==1.12
tifffile==2023.12.9
tokenizers==0.15.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
transformers==4.36.2
typing_extensions==4.9.0
urllib3==2.0.7
uvicorn==0.25.0
